% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_draw.R
\name{eeval_draw}
\alias{eeval_draw}
\title{Funkcja rysuje wykresy na podstawie danych z obliczen eeval_calc()}
\usage{
eeval_draw(
  dane = eeval_wynik,
  x = .data$Nat,
  y = .data$Emisja,
  z = .data$Pollutant,
  u = .data$Euro.Standard,
  nrow = 2,
  skala_y = 3000
)
}
\arguments{
\item{dane}{data.frame}

\item{x}{character/numeric}

\item{y}{character}

\item{z}{character}

\item{u}{character}

\item{nrow}{numeric}

\item{skala_y}{numeric}
}
\value{
double
}
\description{
Funkcja rysuje wykresy na podstawie danych z obliczen eeval_calc()
}
\examples{
eeval_draw()

}
