% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_wynik.R
\docType{data}
\name{eeval_wynik}
\alias{eeval_wynik}
\title{Wynik obliczen emisji dla przykladowego pliku input}
\format{
Ramka danych zawiera:
\describe{
\item{Category}{Jak we wskazniki}
\item{Fuel}{Jak we wskazniki}
\item{Euro.Standard}{Jak we wskazniki}
\item{Technology}{Jak we wskazniki}
\item{Pollutant}{Jak we wskazniki}
\item{Mode}{Jak we wskazniki}
\item{Segment}{Jak we wskazniki}
\item{Nat}{Jak we wskazniki}
\item{Emisja}{Jak we wskazniki}
}
}
\usage{
eeval_wynik
}
\description{
Wynik obliczen emisji dla przykladowego pliku input
}
\examples{
eeval_wynik
}
\keyword{datasets}
