% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Dane wejsciowe dla funkcji szacowania emisji}
\format{
Ramka danych zawiera:
\describe{
\item{Nat}{Natezenie ruchu}
\item{Segment}{Segment pojazdow - Mini itd.}
\item{Fuel}{Rodzaj paliwa - Petrol itd.}
\item{Technology}{Technologia pracy silnika - GDI itd.}
}
}
\usage{
input
}
\description{
Natezenie ruchu to wygenerowany losowo wektor o zadanej dlugosci.
Pozostale dane sa wybrane z obiektu wskazniki.
}
\examples{
input
}
\keyword{datasets}
