% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Dane potrzebne dla funkcji szacowania emisji}
\format{
Ramka danych zawiera:
\describe{
\item{Category}{Passenger Cars itd.}
\item{Fuel}{Petrol itd.}
\item{Segment}{Mini itd.}
\item{Euro.Standard}{Euro 4 itd.}
\item{Technology}{GDI itd.}
\item{Pollutant}{NOx itd.}
\item{Mode}{Urban Peak itd.}
\item{Alpha}{double}
\item{Beta}{double}
\item{Gamma}{double}
\item{Delta}{double}
\item{Epsilon}{double}
\item{Zita}{double}
\item{Hta}{double}
\item{Reduction}{int}
\item{Bio}{int}
\item{Procent}{int}
}
}
\usage{
wskazniki
}
\description{
Dane potrzebne dla funkcji szacowania emisji
}
\examples{
 wskazniki
}
\keyword{datasets}
