% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval.R
\docType{package}
\name{eeval}
\alias{eeval}
\title{eeval: Pakiet oceny emisji (emission evaluation)}
\description{
Pakiet służy do obliczania emisji zanieczyszczeń generowanych przez pojazdy
z silnikami spalinowymi i wyswietlania wyników na wykresach.
}
\details{
Pakiet zawiera trzy funkcje: eeval_input(), eeval_calc(), eeval_draw()

\code{eeval_input()} generuje dane wejsciowe

\code{eeval_calc()} oblicza emisje

\code{eeval_draw()} rysuje wykresy emisji
}
\section{Dane do obliczen}{

Funkcje pakietu korzystaja z danych dolaczonych do pakietu.

Dane sa dostepne dla uzytkownika jako obiekty \code{wskazniki}, \code{input}, \code{wynik_eeval}.
}

