% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval.R
\docType{package}
\name{eeval}
\alias{eeval}
\title{eeval: Pakiet oceny emisji (emission evaluation)}
\description{
Pakiet zawiera dwie funkcje:
eeval_calc oblicza emisje
eeval_draw rysuje wykresy emisji
}
\section{Dane do obliczen}{

Funkcje pakietu korzystaja z danych dolaczonych do pakietu.
Dane sa dostepne dla uzytkownika jako obiekty wskazniki i input.
}

