% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_calc.R
\name{eeval_calc}
\alias{eeval_calc}
\title{Funkcja do wyznaczania wartosci emisji}
\usage{
eeval_calc(
  dane = input,
  kategoria = "Passenger Cars",
  euro = c("Euro 3", "Euro 4", "Euro 5", "Euro 6 up to 2016"),
  mode = "",
  substancja = c("EC", "CO", "NOx")
)
}
\arguments{
\item{dane}{data.frame - dane wejsciowe w formacie jak w obiekcie input}

\item{kategoria}{character - kategoria pojazdu - Passenger Cars itd.}

\item{euro}{character - norma dopuszczalnych emisji spalin - Euro 4 itd.}

\item{mode}{character - tryb jazdy pojazdu - Urban Peak itd.}

\item{substancja}{character - rodzaj emitowanego zanieczyszczenia - NOx itd.}
}
\value{
double
}
\description{
Funkcja oblicza emisje na podstawie danych zawartych w zmiennej
\code{wskazniki} oraz przygotowanej w odpowiedni sposob zmiennej \code{input}.
W obiekcie \code{wskazniki} zapisane sa dane emisji z podzialem na rozne kategorie.
W obiekcie \code{input} zapisane sa wygenerowane losowo dane natezenia ruchu.
}
\details{
Algorytm obliczen emisji:
Emisja = Nat x ((Alpha  x Procent^2 + Beta x Procent + Gamma +
(Delta/Procent)) / (Epsilon x Procent^2 + Zita x Procent + Hta) x (1- Reduction))
}
\examples{
# Obliczenie emisji dla parametrow domyslnych (input)
  eeval_wynik <- eeval_calc()

\dontrun{
# Obliczenie emisji dla przykladowych danych input2
eeval_wynik2 <-
 eeval_calc(dane = input2,
            kategoria = "Passenger Cars",
            euro = c("Euro 3", "Euro 4", "Euro 5", "Euro 6 up to 2016"),
            substancja = c("EC", "CO", "NOx"),
            mode = "")

# Obliczenie emisji dla przykladowych danych inut3
eeval_wynik3 <-
 eeval_calc(dane = input3,
            kategoria = "Buses",
            euro = c("Euro II", "Euro III", "Euro IV", "Euro V"),
            substancja = c("EC", "CO", "NOx", "NH3"),
            mode = "")
}
}
