% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_calc.R
\name{eeval_calc}
\alias{eeval_calc}
\title{Funkcja oblicza wartosc emisji (wersja v3)}
\usage{
eeval_calc(
  dane = input,
  kategoria = "Passenger Cars",
  euro = c("Euro 3", "Euro 4", "Euro 5", "Euro 6 up to 2016"),
  mode = "",
  substancja = c("EC", "CO", "NOx")
)
}
\arguments{
\item{dane}{data.frame}

\item{kategoria}{character}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{
double
}
\description{
Argumentami sa norma emisji spalin i rodzaj substancji (wskaznika)
}
\examples{
eeval_wynik <- eeval_calc()

}
