% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_calc.R
\name{eeval_calc}
\alias{eeval_calc}
\title{Funkcja oblicza wartosc emisji (wersja v2)}
\usage{
eeval_calc(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{dane}{data.frame}

\item{kategoria}{character}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{
double
}
\description{
Argumentami sa norma emisji spalin i rodzaj substancji (wskaznika)
}
\examples{
eeval_calc(euro = "Euro 5", substancja = c("EC", "CO", "NOx"))

}
