% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_calc.R
\name{eeval_calc}
\alias{eeval_calc}
\title{Funkcja mnozy dwa argumrnty}
\usage{
eeval_calc(x, y)
}
\arguments{
\item{x}{double}

\item{y}{double}
}
\value{
double
}
\description{
Funkcja mnozy dwa argumrnty
}
\examples{
eeval_calc(2, 4)
}
