% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_draw.R
\name{eeval_draw}
\alias{eeval_draw}
\title{Funkcja rysuje wykresy na podstawie danych utworzonych za pomoca eeval_calc()}
\usage{
eeval_draw(
  dane = eeval_wynik,
  x = .data$Nat,
  y = .data$Emisja,
  z = .data$Pollutant,
  u = .data$Euro.Standard,
  nrow = 2,
  max_y = 3000
)
}
\arguments{
\item{dane}{data.frame}

\item{x}{character lub numeric - kolumna ramki danych na osi x}

\item{y}{numeric - kolumna ramki danych na osi y}

\item{z}{character - dane dla kolejnych wykresow}

\item{u}{character - dane dla kolejnych wykresow}

\item{nrow}{liczba wykresow w jednym wierszu}

\item{max_y}{maksymalna wartosc na osi y}
}
\value{
double
}
\description{
Rodzaj wykresow jest ustalany automatycznie w zaleznosci od typu
zmiennej na osi x.
Dla zmiennej typu numeric rysowane sa wykresy punktowe z wygladzonymi liniami.
Dla zmiennej typu character rysowane sa wykresy pudelkowe.
}
\examples{
eeval_draw()

}
