% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_input.R
\name{eeval_input}
\alias{eeval_input}
\title{Funkcja przygotowuje danedla funkcji eeval_calc() (wersja v2)}
\usage{
eeval_input(
  segm = c("Mini", "Small", "Medium", "Large-SUV-Executive"),
  n_spl = 300
)
}
\arguments{
\item{segm}{wektor wartosci Segment dla odpowiadajacej pozycji Category}

\item{n_spl}{double}
}
\value{
data.frame
}
\description{
Funkcja przygotowuje danedla funkcji eeval_calc() (wersja v2)
}
\examples{
input2 <- eeval_input()

}
