% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeval_wynik.R
\docType{data}
\name{eeval_wynik}
\alias{eeval_wynik}
\title{Wynik obliczen emisji dla przykladowego pliku input}
\format{
Ramka danych zawiera:
\describe{
\item{Category}{Kategoria pojazdu - Passenger Cars itd.}
\item{Fuel}{Rodzaj paliwa - Petrol itd.}
\item{Euro.Standard}{Norma dopuszczalnych emisji spalin - Euro 4 itd.}
\item{Technology}{Technologia pracy silnika - GDI itd.}
\item{Pollutant}{Rodzaj emitowanego zanieczyszczenia - NOx itd.}
\item{Mode}{Tryb jazdy pojazdu - Urban Peak itd.}
\item{Segment}{Segment pojazdow - Mini itd.}
\item{Nat}{Natezenie ruchu - wygenerowany losowo wektor o zadanej dlugosci}
\item{Emisja}{Obliczona emisja spalin}
}
}
\usage{
eeval_wynik
}
\description{
Obiekt ten stanowi przykladowe dane wejsciowe dla funkcji eeval_draw().
}
\examples{
eeval_wynik
}
\keyword{datasets}
