% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input}
\alias{input}
\title{Dane potrzebne dla funkcji szacowania emisji}
\format{
Ramka danych zawiera:
\describe{
\item{Nat}{Natezenie ruchu}
\item{Segment}{Jak we wskazniki}
\item{Fuel}{Jak we wskazniki}
\item{Technology}{Jak we wskazniki}
}
}
\source{
Wylosowane dane
}
\usage{
input
}
\description{
Dane potrzebne dla funkcji szacowania emisji
}
\examples{
\dontrun{
 input
 }
}
\keyword{datasets}
