% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Dane pomocnicze dla funkcji szacowania emisji}
\format{
Ramka danych zawiera:
\describe{
\item{Nat}{Nat = abs(rnorm(50, mean = 100, sd = 50))}
\item{Segment}{sample(c("Mini","Small","Medium","Large-SUV-Executive"), ...}
\item{Fuel}{sample(unique(wskazniki$Fuel), ...}
\item{Technology}{sample(unique(wskazniki$Technology), ...}
}
}
\usage{
input
}
\description{
Z obiektu wskaznik
zostało wylosowane 50 wierszy i zapisane do pliku input.rda
}
\examples{
input
}
\keyword{datasets}
