% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Dane potrzebne dla funkcji szacowania emisji}
\format{
Ramka danych zawiera:
\describe{
\item{Category}{Passenger Cars itd.}
\item{Fuel}{Petrol itd.}
}
}
\source{
Plik emisja2019.xlsx
}
\usage{
wskazniki
}
\description{
Dane potrzebne dla funkcji szacowania emisji
}
\examples{
\dontrun{
 wskazniki
}
}
\keyword{datasets}
